requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Limiter==3.9.2
Flask-Migrate==4.0.7
email_validator==2.2.0